Update Spark cluster launch scripts

+++ containers/singularity/stop_spark_cluster.sh
@@ -1,11 +1,9 @@
 #!/bin/bash
 
-num_workers=${1}
-
-rm master_success
-
-wait
-
-sleep 10
-
-./stop_spark_master.sh
+if [ ${PMIX_RANK} -eq "0" ]; then
+    sleep 10
+    ./stop_spark_master.sh
+    rm master_success
+else
+    singularity instance stop spark-process
+fi
